+++ docs/conf.py
@@ -56,6 +56,10 @@ extensions = [
     'sphinx.ext.viewcode',   # disabed, raises anexception
     'sphinx.ext.ifconfig',
     'sphinx_git',            # requires 'sphinx-git' Python package
+    # Note: sphinx_rtd_theme is not compatible with sphinxcontrib.fulltoc,
+    # but since it already provides a full TOC in the navigation pane, the
+    # sphinxcontrib.fulltoc extension is not needed.
+    'sphinx_rtd_theme',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -196,13 +200,17 @@ viewcode_import = True
 
 # The theme to use for HTML and HTML Help pages.
 # See http://www.sphinx-doc.org/en/stable/theming.html for built-in themes.
-html_theme = "classic"
+html_theme = 'sphinx_rtd_theme'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.
 # See http://www.sphinx-doc.org/en/stable/theming.html for the options
 # available for built-in themes.
+# For options of the 'sphinx_rtd_theme', see
+# https://sphinx-rtd-theme.readthedocs.io/en/latest/configuring.html
 html_theme_options = {
+    'style_external_links': False,
+    'collapse_navigation': False,
 }
 
 # Add any paths that contain custom themes here, relative to this directory.
@@ -437,6 +445,8 @@ intersphinx_mapping = {
   'py': ('https://docs.python.org/2/', None), # agnostic to Python version
   'py2': ('https://docs.python.org/2', None), # specific to Python 2
   'py3': ('https://docs.python.org/3', None), # specific to Python 3
+  'pywbem': ('https://pywbem.readthedocs.io/en/latest', None),
+  # TODO: Go back to using stable once pywbem 1.1.0 is released.
 }
 
 intersphinx_cache_limit = 5
@@ -483,19 +493,4 @@ intersphinx_cache_limit = 5
 extlinks = {
   'nbview': ('http://nbviewer.jupyter.org/github/pywbem/pywbemcli/blob/master/docs/notebooks/%s', ''),
   'nbdown': ('https://github.com/pywbem/pywbemcli/raw/master/docs/notebooks/%s', '')
-}
-
-
-# http://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html
-# Add the "intersphinx" extension
-# Add fulltoc to show the chapter TOC in the sidebar for the pywbemcli section
-# Also always shows the chapters in the sidebar
-extensions = [
-    'sphinx.ext.intersphinx',
-    'sphinxcontrib.fulltoc',
-]
-# Addintersphinx mappings
-intersphinx_mapping = {
-    'pywbem': ('https://pywbem.readthedocs.io/en/latest', None),
-    # TODO: Go back to using stable once pywbem 1.1.0 is released.
 }